Extract conversion code to Parmesan::Converter
