+++ lib/parmesan/parameter.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'set'
+require_relative 'converter'
 
 module Parmesan
   class Parameter
@@ -59,7 +60,7 @@ module Parmesan
 
     def value(request)
       value = collect(request)
-      convert_top_level(value)
+      convert(value)
     end
 
     def collect(request)
@@ -67,6 +68,10 @@ module Parmesan
       return collect_object(request) if object?
 
       Rack::Utils.parse_query(request.query_string)[name]
+    end
+
+    def convert(value)
+      Converter.call(value, schema)
     end
 
     def collect_array(request)
@@ -97,31 +102,6 @@ module Parmesan
 
     def object?
       type == 'object'
-    end
-
-    def convert_top_level(value)
-      return convert_object(value) if object?
-
-      convert(schema, value)
-    end
-
-    def convert(schema, value)
-      case schema && schema['type']
-      when 'integer'
-        Integer(value, 10)
-      when 'number'
-        Float(value)
-      when 'boolean'
-        value == 'true'
-      else
-        value
-      end
-    end
-
-    def convert_object(value)
-      value.each_with_object({}) do |(k, v), hsh|
-        hsh[k] = convert(schema.fetch('properties').fetch(k), v)
-      end
     end
 
     def ref?(object)
